lego_measure/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
flask>=3.0.0
